refactor: move adapter-specific errors to their own packages
This refactor improves package separation by moving adapter-specific error classes from the core @tanstack/db package to their respective adapter packages.

Changes:
- Moved Electric DB errors to @tanstack/electric-db-collection/errors
- Moved TrailBase DB errors to @tanstack/trailbase-db-collection/errors
- Moved Query Collection errors to @tanstack/query-db-collection/errors
- Updated imports in adapter files to use local error classes
- Kept only generic/core errors in @tanstack/db
- Updated changeset to reflect new package structure
- Fixed linting issues (shadow variables and unnecessary conditions)

Benefits:
- Better package separation and encapsulation
- Reduced core package size
- Error classes are co-located with their usage
- Each adapter manages its own error types
