+++ packages/db/src/errors.ts
@@ -516,114 +516,6 @@ export class InvalidStorageObjectFormatError extends LocalStorageCollectionError
   }
 }
 
-// Electric DB Collection Errors
-export class ElectricDBCollectionError extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = `ElectricDBCollectionError`
-  }
-}
-
-export class ExpectedNumberInAwaitTxIdError extends ElectricDBCollectionError {
-  constructor(txIdType: string) {
-    super(`Expected number in awaitTxId, received ${txIdType}`)
-  }
-}
-
-export class TimeoutWaitingForTxIdError extends ElectricDBCollectionError {
-  constructor(txId: number) {
-    super(`Timeout waiting for txId: ${txId}`)
-  }
-}
-
-export class ElectricInsertHandlerMustReturnTxIdError extends ElectricDBCollectionError {
-  constructor() {
-    super(
-      `Electric collection onInsert handler must return a txid or array of txids`
-    )
-  }
-}
-
-export class ElectricUpdateHandlerMustReturnTxIdError extends ElectricDBCollectionError {
-  constructor() {
-    super(
-      `Electric collection onUpdate handler must return a txid or array of txids`
-    )
-  }
-}
-
-export class ElectricDeleteHandlerMustReturnTxIdError extends ElectricDBCollectionError {
-  constructor() {
-    super(
-      `Electric collection onDelete handler must return a txid or array of txids`
-    )
-  }
-}
-
-// TrailBase DB Collection Errors
-export class TrailBaseDBCollectionError extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = `TrailBaseDBCollectionError`
-  }
-}
-
-export class TimeoutWaitingForIdsError extends TrailBaseDBCollectionError {
-  constructor(ids: string) {
-    super(`Timeout waiting for ids: ${ids}`)
-  }
-}
-
-export class ExpectedInsertTypeError extends TrailBaseDBCollectionError {
-  constructor(actualType: string) {
-    super(`Expected 'insert', got: ${actualType}`)
-  }
-}
-
-export class ExpectedUpdateTypeError extends TrailBaseDBCollectionError {
-  constructor(actualType: string) {
-    super(`Expected 'update', got: ${actualType}`)
-  }
-}
-
-export class ExpectedDeleteTypeError extends TrailBaseDBCollectionError {
-  constructor(actualType: string) {
-    super(`Expected 'delete', got: ${actualType}`)
-  }
-}
-
-// Query Collection Errors
-export class QueryCollectionError extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = `QueryCollectionError`
-  }
-}
-
-export class QueryKeyRequiredError extends QueryCollectionError {
-  constructor() {
-    super(`[QueryCollection] queryKey must be provided.`)
-  }
-}
-
-export class QueryFnRequiredError extends QueryCollectionError {
-  constructor() {
-    super(`[QueryCollection] queryFn must be provided.`)
-  }
-}
-
-export class QueryClientRequiredError extends QueryCollectionError {
-  constructor() {
-    super(`[QueryCollection] queryClient must be provided.`)
-  }
-}
-
-export class GetKeyRequiredError extends QueryCollectionError {
-  constructor() {
-    super(`[QueryCollection] getKey must be provided.`)
-  }
-}
-
 // Sync Cleanup Errors
 export class SyncCleanupError extends Error {
   constructor(collectionId: string, error: Error | string) {
