+++ packages/electric-db-collection/src/index.ts
@@ -4,3 +4,5 @@ export {
   type ElectricCollectionUtils,
   type Txid,
 } from "./electric"
+
+export * from "./errors"
