+++ packages/query-db-collection/src/index.ts
@@ -3,3 +3,5 @@ export {
   type QueryCollectionConfig,
   type QueryCollectionUtils,
 } from "./query"
+
+export * from "./errors"
