+++ packages/trailbase-db-collection/src/index.ts
@@ -3,3 +3,5 @@ export {
   type TrailBaseCollectionConfig,
   type TrailBaseCollectionUtils,
 } from "./trailbase"
+
+export * from "./errors"
